+++ aql/aql.proto
@@ -11,16 +11,13 @@ message GraphQuery {
 }
 
 message GraphQuerySet {
-  // TODO a sub query done by match() shouldn't change graphs.
   repeated GraphQuery queries = 1;
 }
 
 message GraphStatement {
     oneof statement {
-        // TODO get rid of ListValue?
         google.protobuf.ListValue V = 1;
-        // TODO why is E only a single string?
-        google.protobuf.ListValue E = 2;
+        string E = 2;
 
         HasStatement has = 5;
         google.protobuf.ListValue hasLabel = 6;
@@ -32,7 +29,6 @@ message GraphStatement {
         google.protobuf.ListValue outEdge = 13;
         google.protobuf.ListValue both = 14;
         google.protobuf.ListValue bothEdge = 15;
-        // TODO breaking protobuf style. need linter
 
         google.protobuf.ListValue outBundle = 16;
 
@@ -43,20 +39,19 @@ message GraphStatement {
         int64 limit = 25;
         string count = 26;
 
-        // TODO add "first"
         string groupCount = 30;
 
         GraphQuerySet match = 40;
 
         //Function Methods
-        // TODO probably doesn't belong as a statement. should be global
         string import = 50;
         string map = 51;
-        string fold = 52;
-        string filter = 53;
-        string filterValues = 54;
+        FoldStatement fold = 52;
+        FoldStatement vertexFold = 53;
+        string filter = 54;
+        string filterValues = 55;
 
-        string vertexFromValues = 55;
+        string vertexFromValues = 56;
     }
 }
 
@@ -67,6 +62,11 @@ message HasStatement {
 
 message SelectStatement {
     repeated string labels = 1;
+}
+
+message FoldStatement {
+  string source = 1;
+  google.protobuf.Value init = 2;
 }
 
 message Vertex {
@@ -93,13 +93,10 @@ message Bundle {
 
 message QueryResult {
     oneof result {
-      google.protobuf.Struct struct = 1;
-      Vertex vertex = 2;
-      Edge edge = 3;
-      Bundle bundle = 4;
-      int32 int_value = 5;
-      double float_value = 6;
-      string str_value = 7;
+      Vertex vertex = 1;
+      Edge edge = 2;
+      Bundle bundle = 3;
+      google.protobuf.Value data = 4;
     }
 }
 
@@ -209,6 +206,7 @@ service Edit {
   rpc AddSubGraph(Graph) returns (EditResult) {
     option (google.api.http) = {
       post: "/v1/graph/{graph}/subgraph"
+      body: "*"
     };
   }
